changed providerId type to long, moved setting the sorted header html class assignment to the view model and added a unit test to test said assignment

+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.Commitments.Shared/Interfaces/ICommitmentsService.cs
@@ -13,6 +13,6 @@ namespace SFA.DAS.Commitments.Shared.Interfaces
         Task<CohortDetails> GetCohortDetail(long cohortId);
         Task<EditDraftApprenticeshipDetails> GetDraftApprenticeshipForCohort(long cohortId, long draftApprenticeshipId);
         Task UpdateDraftApprenticeship(long cohortId, long draftApprenticeshipId, UpdateDraftApprenticeshipRequest updateRequest);
-        Task<GetApprenticeshipsResponse> GetApprenticeships(uint providerId, string sortField = "", bool reverseSort = false);
+        Task<GetApprenticeshipsResponse> GetApprenticeships(long providerId, string sortField = "", bool reverseSort = false);
     }
 }
